+++ src/main/kotlin/org/wfanet/panelmatch/client/exchangetasks/JoinKeyTasksImpl.kt
@@ -16,16 +16,17 @@ package org.wfanet.panelmatch.client.exchangetasks
 
 import com.google.protobuf.ByteString
 import org.wfanet.measurement.api.v2alpha.ExchangeWorkflow
+import org.wfanet.measurement.common.toLocalDate
 import org.wfanet.panelmatch.client.common.ExchangeContext
 import org.wfanet.panelmatch.client.privatemembership.PrivateMembershipCryptor
 import org.wfanet.panelmatch.common.certificates.CertificateManager
 import org.wfanet.panelmatch.common.crypto.DeterministicCommutativeCipher
 
-class GenerateKeysTasksImpl(
+class JoinKeyTasksImpl(
   private val deterministicCommutativeCryptor: DeterministicCommutativeCipher,
   private val getPrivateMembershipCryptor: (ByteString) -> PrivateMembershipCryptor,
   private val certificateManager: CertificateManager
-) : GenerateKeysTasks {
+) : JoinKeyTasks {
   override fun generateLookupKeys(): ExchangeTask {
     return GenerateLookupKeysTask()
   }
@@ -44,5 +45,19 @@ class GenerateKeysTasksImpl(
 
   override fun generateExchangeCertificate(context: ExchangeContext): ExchangeTask {
     return GenerateExchangeCertificateTask(certificateManager, context.exchangeDateKey)
+  }
+
+  override fun intersectAndValidate(context: ExchangeContext): ExchangeTask {
+    val step = context.step
+    check(step.stepCase == ExchangeWorkflow.Step.StepCase.INTERSECT_AND_VALIDATE_STEP)
+
+    val maxSize = step.intersectAndValidateStep.maxSize
+    val maximumNewItemsAllowed = step.intersectAndValidateStep.maximumNewItemsAllowed
+
+    return IntersectValidateTask(
+      maxSize = maxSize,
+      maximumNewItemsAllowed = maximumNewItemsAllowed,
+      isFirstExchange = context.date == context.workflow.firstExchangeDate.toLocalDate()
+    )
   }
 }
